lots of stuff

+++ entity/checkin.go
@@ -5,7 +5,7 @@ import "time"
 type CheckIn struct {
 	StdFields
 
-	UserID      int       `db:"user_id"       json:"user_id"`
+	UserID      int64     `db:"user_id"       json:"user_id"`
 	CheckInTime time.Time `db:"check_in_time" json:"check_in_time"`
 	LastSent    time.Time `db:"last_sent"     json:"last_sent"`
 }
